Refactor: Modulize get random
