+++ app/modules/db.server.ts
@@ -453,3 +453,76 @@ export async function getRecentComments(){
 
     return recentComments;
 }
+
+export async function getRandomPosts(): Promise<PostCardData[]> {
+    /*
+    prisma.dimPosts.findManyRandomを利用してもランダムな記事を取得することは可能であるが、タイムアウトしてしまうため、インデックスを作成したuuidを使って疑似的にランダムな記事を取得している
+    */
+    const postCount = await prisma.dimPosts.count();
+    const randomPostOffset = Math.max(
+        Math.floor(Math.random() * postCount) - 10,
+        0
+    );
+
+    const randomPostsRaw = await prisma.dimPosts.findMany({
+        select: {
+            postId: true,
+            postTitle: true,
+            postDateGmt: true,
+            countLikes: true,
+            countDislikes: true,
+            ogpImageUrl: true,
+            rel_post_tags: {
+                select: {
+                    dimTag: {
+                        select: {
+                            tagName: true,
+                            tagId: true,
+                        },
+                    },
+                },
+            },
+        },
+        orderBy: { uuid : "asc"},
+        skip: randomPostOffset,
+        take: 10,
+    })
+    const commentCount = await prisma.dimComments.groupBy({
+        by: ["postId"],
+        _count: { commentId: true },
+        where: { postId: { in: randomPostsRaw.map((post) => post.postId) } },
+    })
+    const randomPosts = randomPostsRaw.map((post) => {
+        const count = commentCount.find((c) => c.postId === post.postId)?._count.commentId || 0;
+        return {
+            ...post,
+            countComments: count,
+            tags: post.rel_post_tags.map((tag) => tag.dimTag),
+        }
+    })
+    return randomPosts;
+}
+
+export async function getRandomComments(){
+    const commentCount = await prisma.dimComments.count();
+    const randomCommentOffset = Math.max(
+        Math.floor(Math.random() * commentCount) - 10,
+        0
+    );
+
+    const randomComments = await prisma.dimComments.findMany({
+        select: {
+            commentId: true,
+            commentContent: true,
+            commentDateGmt: true,
+            commentAuthor: true,
+            postId: true,
+            dimPosts: { select: { postTitle: true } },
+        },
+        orderBy: { uuid: "asc" },
+        skip: randomCommentOffset,
+        take: 10,
+    })
+    return randomComments;
+}
+
